fix: E0609 missing field errors + unclosed delimiter
- Fix unclosed delimiter in handlers_consciousness.rs (add closing brace for impl block)
- Fix E0609: b.capability → b.brain.capability (SelfIteratingBrain has brain field)
- Fix E0609: d.ts_nanos → d["ts_nanos"] (JSON field access)
- Fix E0609: AgentInfo fields accessed through config struct
- Fix E0609: Fingerprint only has user_agent/platform/language/viewport/timezone

+++ neotrix-core/src/l5_cognition/nt_core/nt_core_resonance_complexity.rs
@@ -172,7 +172,7 @@ impl ResonanceComplexityEngine {
         let g = 2.0; // 信号增益
         let tau = 150.0; // 驻留时间 (ms)
 
-        let composite: f64 = (d * c * g * tau).sqrt() / 100.0_f64;
+        let composite: f64 = (d as f64 * c as f64 * g as f64 * tau as f64).sqrt() / 100.0_f64;
 
         ComplexityIndex {
             fractal_dimension: d,
